test: ensure that we can run 'rake minitest' on minitest v5
